Fix button state of second page.
When sex and marital status have not been entered, the bottom of the screen shows microphone button instead of the "Edit" button.

Also require at least one medical condition to be selected before continuing.

This improves the user flow the app by making the microphone button appear at the start of every page.